+++ src/packages/piral-cli/src/bundler.ts
@@ -20,19 +20,43 @@ const bundlers: Array<QualifiedBundler> = [];
 async function installDefaultBundler(root: string) {
   const parcel = 'piral-cli-parcel';
 
-  try {
-    log('generalDebug_0003', `Trying to resolve ${parcel}.`);
-    require(parcel);
-  } catch {
-    log('generalDebug_0003', `Determining NPM client at "${root}" ...`);
-    const client = await determineNpmClient(root);
-    log('generalDebug_0003', `Prepare to install ${parcel}@${cliVersion} using "${client}" into "${root}".`);
-    progress(`Installing ${parcel} ...`);
-    await installPackage(client, `${parcel}@${cliVersion}`, root, '--save-dev');
-    log('generalDebug_0003', 'Installed bundler.');
-  }
+  // try {
+  //   log('generalDebug_0003', `Trying to resolve ${parcel}.`);
+  //   require(parcel);
+  // } catch {
+  // -- that did not work out as expected, since failed module
+  // -- requires seem to be captured in the context.
+  // -- instead, we'll assume an installation is necessary
+  // }
+
+  log('generalDebug_0003', `Determining NPM client at "${root}" ...`);
+  const client = await determineNpmClient(root);
+  log('generalDebug_0003', `Prepare to install ${parcel}@${cliVersion} using "${client}" into "${root}".`);
+  progress(`Installing ${parcel} ...`);
+  await installPackage(client, `${parcel}@${cliVersion}`, root, '--save-dev');
+  log('generalDebug_0003', 'Installed bundler.');
 
   require('./inject').inject(parcel);
+}
+
+function checkDefaultBundler(bundler: QualifiedBundler) {
+  if (!bundler?.actions) {
+    fail('defaultBundlerMissing_0073');
+  }
+
+  return bundler;
+}
+
+function checkCustomBundler(bundler: QualifiedBundler, bundlerName: string) {
+  if (!bundler?.actions) {
+    fail(
+      'bundlerMissing_0072',
+      bundlerName,
+      bundlers.map(b => b.name),
+    );
+  }
+
+  return bundler;
 }
 
 async function findBundler(root: string, bundlerName?: string) {
@@ -40,19 +64,11 @@ async function findBundler(root: string, bundlerName?: string) {
 
   if (bundlerName) {
     const [bundler] = bundlers.filter(m => m.name === bundlerName);
-
-    if (!bundler) {
-      fail(
-        'bundlerMissing_0072',
-        bundlerName,
-        bundlers.map(b => b.name),
-      );
-    }
-
-    return bundler;
+    return checkCustomBundler(bundler, bundlerName);
   } else if (!defaultBundler) {
     await installDefaultBundler(root);
-    return bundlers[0];
+    const [bundler] = bundlers;
+    return checkDefaultBundler(bundler);
   } else {
     return defaultBundler;
   }